Option to position mix nodes halfway, or bottom
Previously it would always position the new mix nodes (from operators
like Ctrl+Numpad0) at the bottom in line with the bottom-most node. For
simple cases such as mixing only two nodes, it's better looking to place
the new node halfway between the two.

This should be improved to use the dimensions of the objects (looks ugly
in cases such as the large Mapping node, though this won't be used often
I think)

Also, an option to have the final merge node (in cases where mixing many
nodes at once) is put at the top, rather than the bottom, would be nice.