Replaced built-in Make-link fnc with own
Used in the Lazy Connect operator - this just checks if a socket was
already connected and then try the next one instead. Also opens this up
to future improvements and guesses.

Also replaced node-space mouse position hack with a proper method.

And re-added hotkey for output swapping as Alt+S