Added 'NW' prefix to class names
I don't know enough about python or coding to be sure if this is
necessary, but might as well just to be safe.

Also changed old operator namespace to 'nw.opname' instead of
'node.opname' to keep all the addon's operators in one place and not
mixed with original code.